Fix qr code decoding
It turns out that reading and concatinating the "SCAN_RESULT_BYTE_SEGMENTS_{x}" within Utils.readSQRLQRCode would sometimes leave us with chunks of the SQRL login URL missing. So instead, we will now use the getContents() method of the parsed IntentResult if we're dealing with login QR codes, which seems to always deliver the full URL. Only for the binary identity qr codes, we are still forced to use the raw byte segments.